refactor(v4.0)!: consolidate filter pipeline into filter/ package
Sprint 1 of the v4.0 architecture refactor. Collapses three modules and
three Pydantic models that previously implemented one logical concern
(build + validate + parse + explain a ServiceNow filter) into a single
cohesive package.

Layout:
    filter/
      __init__.py      # public API re-exports
      models.py        # TableFilterParams, SmartQueryParams,
                       # QueryValidationResult
      builder.py       # ServiceNowQueryBuilder (moved from
                       # query_validation.py)
      validator.py     # validate_*_filter funcs + validate_and_correct_filters
                       # (new — owns the auto-correction logic that
                       # previously lived in QueryIntelligence)
      intelligence.py  # QueryIntelligence (moved from query_intelligence.py)
      explainer.py     # QueryExplainer, explain_existing_filter

Backref break: filter/intelligence.py no longer touches filter/builder.py.
The NL parser used to call ServiceNowQueryBuilder.build_priority_or_filter
directly to auto-correct comma syntax; that logic now lives in the
validator (validate_and_correct_filters), which is the only module
allowed to bridge intelligence -> builder. Public API of
QueryIntelligence._validate_and_improve_filters is preserved as a thin
delegate so mock patches keep working.

Pydantic models relocation: TableFilterParams and SmartQueryParams moved
from Table_Tools/generic_table_tools.py to filter/models.py. The class
names are re-exported from generic_table_tools so downstream imports
(consolidated_tools, intelligent_query_tools, etc.) keep working
without churn. pydantic.BaseModel/Field no longer imported in
generic_table_tools.

Backwards compat: query_validation.py and query_intelligence.py become
thin re-export shims. All existing
  from query_validation import ServiceNowQueryBuilder
  from query_intelligence import QueryIntelligence
call sites — including patch targets like
  @patch('query_intelligence.extract_keywords')
  @patch('query_validation.validate_query_filters')
— continue to resolve. Shims marked for deletion in v4.1.

Coverage:
- filter/ package: 98.16% line + branch (548 stmts)
- Overall: 82.44%
- 562 tests pass, 5 skipped

Token gate: no read-path behaviour changes — only relocation. Existing
test_token_footprint.py budgets still hold.
